+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Added:
 
 Changed:
 - Devices can now also be mounted on other devices
+- Increased number of workers for gunicorn
 
 ## 1.7.0 - 2023-11-06